PatrykKotlowski/server/warehouse/custom_permissions.py
from rest_framework import permissions
from rest_framework.generics import get_object_or_404


class IsUserAllowed(permissions.BasePermission):
    """
    Object-level permission to only allow owners of an object to edit it.
    Assumes the model instance has an `owner` attribute.
    """

    def has_permission(self, request, view):
        # Read permissions are allowed to any request,
        # so we'll always allow GET, HEAD or OPTIONS requests.

        # Instance must have an attribute named `owner`.
        return request.user.job_title == "Mgr"
