PatrykKotlowski/server/requirements.txt
django
djangorestframework
markdown
django-filter
django-rest-framework-social-oauth2
django-rest-auth
django-cors-headers
dictdiffer